requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-WTF==1.1.1
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
